Maintenance: rewrote hwpch.h into pch.h

+++ Holloware/src/Holloware/Debug/Instrumentor.h
@@ -14,11 +14,11 @@
 //
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <chrono>
 #include <algorithm>
 #include <fstream>
-
 #include <thread>
 
 namespace Holloware
@@ -125,7 +125,7 @@ namespace Holloware
             long long start = std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch().count();
             long long end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
 
-            uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+            uint32_t threadID = (uint32_t)std::hash<std::thread::id>{}(std::this_thread::get_id());
             Instrumentor::Get().WriteProfile({ m_Name, start, end, threadID });
 
             m_Stopped = true;
