+++ Holloware/src/Holloware/ImGui/ImGuiLayer.cpp
@@ -1,15 +1,17 @@
-#include "hwpch.h"
+#include "pch.h"
 #include "ImGuiLayer.h"
 
 #include "Holloware/Core/Window.h"
 #include "Holloware/Core/Application.h"
+#include "Holloware/Core/Layer.h"
+#include "Holloware/Debug/Instrumentor.h"
 #include "Holloware/Events/ApplicationEvent.h"
 #include "Holloware/Events/KeyEvent.h"
 #include "Holloware/Events/MouseEvent.h"
 
-#include "imgui.h"
-#include "backends/imgui_impl_opengl3.h"
-#include "backends/imgui_impl_glfw.h"
+#include <imgui.h>
+#include <backends/imgui_impl_opengl3.h>
+#include <backends/imgui_impl_glfw.h>
 #include <GLFW/glfw3.h>
 
 namespace Holloware
@@ -86,7 +88,7 @@ namespace Holloware
 
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 		// Rendering
 		ImGui::Render();
